moved functions away from misc. refactored import
